Microservices Producer Application
SpringCloudStreamKafkaOrdersourceApplication.Java file, which is actually producing messages.
Supplier<> is a Java function data type. Because there is only one @Bean method that returns this type, Spring Cloud Stream will call this method bean. By default, it will trigger this function once every second and send the result to the default MessageChannel named supply-out-0 (which is constructed on the basis of method name followed by -out-0). But in our case as we have defined the destination topic using @InboundChannelAdapter(channel = Source.OUTPUT), so the messages will be sent to output channel.